app/core/config.py
import os
from dotenv import load_dotenv

# .env 파일에서 환경 변수를 로드합니다.
load_dotenv()

# OpenAI API 키
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

# FAISS 인덱스 경로
FAISS_INDEX_PATH = "faiss_index"
